requirements.txt
diffusers
transformers
accelerate
safetensors
TTS
piper-tts
whisperx
torch==2.0.1
torchaudio==2.0.2